Renamed recoverFeatueFlag.ts to recoverFeatureFlag.ts
